# Decode small maps as PAM, auto-grow to PHM