Start decoding maps into a `TransientArrayMap` initially and allow its assoc method to auto-promote it to a `TransientHashMap` if its size gets large enough. This is consistent with how `(into {} xs)` behaves.