@favila 2018-05-01
That will teach me to make a patch using git's editor...

Note I had to force the JVM timezone to UTC to get the tests to pass. I can make a PR for that if interested (I added `-Duser.timezone=UTC` to `:jvm-opts` in the `project.clj`).