@favila 2018-05-01
Quick decoding speed perf comparison using `lein perf repl` and `(jsonista.json-perf-test/decode-perf)`

Master:
```######################
## decode: jsonista ##
######################

Evaluation count : 1498638 in 6 samples of 249773 calls.
             Execution time mean : 405.190155 ns
    Execution time std-deviation : 12.287989 ns
   Execution time lower quantile : 392.939313 ns ( 2.5%)
   Execution time upper quantile : 417.862655 ns (97.5%)
                   Overhead used : 1.496308 ns```

With PR:

```######################
## decode: jsonista ##
######################

Evaluation count : 1740828 in 6 samples of 290138 calls.
             Execution time mean : 347.886539 ns
    Execution time std-deviation : 7.886340 ns
   Execution time lower quantile : 338.432356 ns ( 2.5%)
   Execution time upper quantile : 356.021850 ns (97.5%)
                   Overhead used : 1.498982 ns```