@ikitommi 2018-05-02
Thanks! I'll run the tests to get new numbers. Curious anoit the timezone setting, haven't seen it myself.